window is resized upon switching

+++ result.h
@@ -2,14 +2,14 @@
 #define RESULT_H
 
 typedef struct {
-	float wpm, time_taken;
+	double wpm, time_taken;
 } Result;
 
-void result_init(Result **, float, float);
+void result_init(Result **, double, double);
 void result_free(Result *);
 
 void
-result_init(Result **result, float wpm, float time_taken) {
+result_init(Result **result, double wpm, double time_taken) {
 	*result = (Result *)malloc(sizeof(Result));
 	(*result)->wpm = wpm;
 	(*result)->time_taken = time_taken;
